move GrowableBuffer header under awkward and add a interface path to it
