+++ include/awkward/builder/BoolBuilder.h
@@ -7,7 +7,7 @@
 #include <string>
 
 #include "awkward/common.h"
-#include "../src/awkward/_v2/cpp-headers/GrowableBuffer.h"
+#include "awkward/GrowableBuffer.h"
 #include "awkward/builder/Builder.h"
 
 namespace awkward {
