+++ include/awkward/builder/Complex128Builder.h
@@ -4,7 +4,7 @@
 #define AWKWARD_COMPLEX128BUILDER_H_
 
 #include "awkward/common.h"
-#include "../src/awkward/_v2/cpp-headers/GrowableBuffer.h"
+#include "awkward/GrowableBuffer.h"
 #include "awkward/builder/Builder.h"
 
 #include <complex>
