+++ include/awkward/builder/Float64Builder.h
@@ -6,7 +6,7 @@
 #include <string>
 
 #include "awkward/common.h"
-#include "../src/awkward/_v2/cpp-headers/GrowableBuffer.h"
+#include "awkward/GrowableBuffer.h"
 #include "awkward/builder/Builder.h"
 
 namespace awkward {
